PyPy version

+++ docker_build.sh
@@ -3,8 +3,19 @@
 NAME="pickapp/pyportredirector"
 VERSION=0.1
 
+[ "$1" == "pypy" ] && VERSION="$VERSION-pypy"
+
 docker rmi -f ${NAME}:${VERSION} 2>/dev/null
-docker rmi -f ${NAME}:latest 2>/dev/null
-docker build --rm -t ${NAME}:${VERSION} -t ${NAME}:latest .
-docker push ${NAME}:${VERSION}
-docker push ${NAME}:latest
+
+if [ "$1" != "pypy" ]; then
+    docker rmi -f ${NAME}:latest 2>/dev/null
+    docker build --rm -t ${NAME}:${VERSION} -t ${NAME}:latest . \
+        && docker push ${NAME}:latest
+fi
+if [ "$1" == "pypy" ]; then
+    docker rmi -f ${NAME}:pypy 2>/dev/null
+    docker build -f Dockerfile.pypy --rm -t ${NAME}:${VERSION} -t ${NAME}:pypy . \
+        && docker push ${NAME}:pypy
+fi
+
+[ $? -eq 0 ] && docker push ${NAME}:${VERSION}
